Problem: cli cannot set fields using project xml
Much of the dss setup is designed for
automation but it is useful to be able
to set a field manually, for example
a small amount of items that do not
merit automated rules, testing or
prototyping. However, rather than
always inputting from scratch it would
be nice to have access to various
materials created in the dss-ui or
other programs and stored as xml.

Solution: combine working keys (glworbs or
other keys to set fields in), project xml
(containing category, part and region names)
and a somewhat simple keyboard-based system
for selecting and setting values (from xml) on
the working keys.